feat: Display more than one line in comment preview
Instead of displaying only the first line of the comment we now display
the full comment, but clipped to a max height.

If the comment contains images, they are displayed as thumbnails.

If user clicks on the expand icon, the modal display the full comment,
with images in their original size.

Part of story #22226: edit test execution comments

Change-Id: If62b168a79edc7780faa9048e07338c5b1a09551