Assume Tuleap is always reached over HTTPS
The sys_https_host setting is removed as it is no more used.

It should not change a lot of things:
 * we already upgrade insecure requests in the nginx configuration by
   default
 * the default setup set the same value for sys_https_host and
   sys_default_domain
 * the `upgrade-insecure-requests` CSP directive already tell browsers
   to upgrade the connection when they can if needed
 * we already uses sys_https_host without verification in few spots

The only instances that could encounter an issue are the ones that had
an empty sys_https_host value with no HTTPS setup or a different value
in sys_default_domain and sys_https_host.

Part of request #22669: Allow to disable HTTP->HTTPS redirection
Closes request #22710: Login forms uses different action URLs depending on the page

Change-Id: I701308fa625cec69553a920ff07248bdb1bb1149