request #22691: Editing a changeset comment fallbacks on the synchronous notification system
The issue is caused by the fact we log the entries in a table indexed by
a changeset ID and the changeset ID stay the same when we edit a
comment.

We log those events in order to warn site administrators if we detect
something that does not look correct but it is outdated. The tracker
plugin is not the only one using the async queue (the pull request
plugin does too for example).

The mecanism specific to the tracker plugin is removed by this contribution
(solving the initial issue) and something more generic working for every
usages of the async queue replaces it.

Change-Id: Ia38b75dee1da716306fa89eb7664a1f3b1aa3e4f