Fix: Use "Switch to…" for SwitchTo button label
The label of the button "Projects…" is confusing:

* When we want to create a project, we have tendency to click on
"Projects" instead of "New"
* When we want to search an artifact, we always have a moment of doubt
before clicking on "Projects"

Mockups of the feature was presenting "Switch to…" as label but we
decided to name it temporarily to "Projects…" in order to not confuse
too much the existing users with the new navigation. With the upcoming
Tuleap 13.0, I suggest that we rename this button.

Documentation screenshots will probably have to be updated, and existing
users might be lost at first sight, but the UX improvement is worth it.

If existing users become completely mad about this change and yell
"REGRESSION!", then a fallback is introduced:

```
tuleap config-set \
  temporarily_keep_legacy_projects_label_for_switchto_menu \
  'I_understand_this_is_a_temporary_configuration_switch_(please_warn_the_Tuleap_dev_team_when_enabling_this)'
```

Closes request #22697: Rename Projects… menu to Switch to…

Change-Id: I7e12199c4951263b3443eef8c09cb54afccdafe5