Clean-up of PossibleParentSelector event handling
/!\ I really hope I got this right, tough review /!\

The initial purpose of this commit was to add tests on the rendering of
the PossibleParentSelector. While doing that, I had to investigate the
various behavious, esp in the most legacy part of that: AgileDashboard.

In `ArtifactParentsSelectorEventListener` the logic is thrown in two
cases:
1. If there is a `child_milestone` parameter
2. If the param `func` is equals to `new-artifact-link`

As far as I can tell, there is no longer any code that produce
`child_milestone` parameter. From my various digging into the code, it
seems to be a remaining part of the Planning v1 when stuff was created
in Tracker and we had redirect between agiledashboard and tracker for
creation. As now everything is handled by the Planning v2 or Taskboard
app the related logic is no longer needed.

For `new-artifact-link`, again AFAICT, it seems to only kick when
someone is creating a new artifact from the Artifact view (aka
lightwindow modal). More specifically, when one tries to create a new
User Story from an Epic artifact, instead of having the parent selector,
the interface displays "Will have $EPIC_TITLE as parent".

However this behavior is already broken because, on master, $EPIC_TITLE
doesnt magically end-up being the parent (probably due to the removal of
the magic of Tracker hierarchy earlier this year). If you don't set the
artifact link type at creation, the user story doesn't end up being a
child.

As the original use case no longer seems called and the only one that is
not dead code no longer actually works, I decided to remove the entire
code.

PS: 👋 to you form the 🛡️ who is gonna read this commit message after
hours of painful debug and back & forth exchanges in the tracker for an
obscure behaviour that caused a REGERESSION! I really meant what I wrote
here, it really seems that this code no longer do anything useful. I
appologize in advance.

Part of story #22681 select a parent from Program for my features/enablers

Change-Id: I0109bea663a44a53e1ed7b94eb2b1dc7a0f60b72