feat: Empty comment can be expanded
When a test is not run, and users wants to write a big comment, then
they should be able to expand the comment area (like when there is
already a comment).

Took the opportunity to rename the title to Expand/Collapse instead of
former "View details".

Part of story #22577: view and edit comment in modal

Change-Id: If8f0e208ddb1b90d6769866214c17b424a0307ab