fix: Propagate pagination info to other plugins
Functionally speaking, this should only affect the
/trackers/:id/parent_artifacts REST route (pagination works again).

The impact of this propagation is larger than initially anticipated.

The main changes are, of course in FeatureDao to take into account the
new parameters and to have the queries for search and count.

Count for pagination was not done, as we usually do with
`SQL_CALC_FOUND_ROWS` because it's actually a bad practice and is
deprecated in MySQL 8.0 https://dev.mysql.com/worklog/task/?id=12615

Part of story #22681 select a parent from Program for my features/enablers

Change-Id: I400b366ae3b1353743e28ae7ce2a919b16aa54ba