feat: display list of attachments of a test execution
When a test execution has attachments (typically dragndrop of an image
in the comment editor, or manual addition of attachments in the
execution artifact ui), they are displayed as a list next to the
comment.

If filename is too long, then ellipsis are displayed to not break the
layout.

Except the download of the corresponding file, no interaction is
available in this list.

The download is opened in a new tab, so that we don't get a warning
about realtime disconnection because SocketIO thinks that we are leaving
the page.

Part of story #22705: display list of attachments of a test execution

Change-Id: I5c8efd46a5f62f9f24c79d0d73279d2dd0710853