Retrieve changeset submission timestamp with interface
part of story #22675 replicate program increment changes to mirrored
program increments

No functional change expected when creating mirror program increments.
You can test for non-regression with the end-to-end tests.

Notes:
In order to create SourceTimeboxChangesetValues for Program Increment
updates, we will need to find the changeset's submission date. We only
have its id. This patch introduces that mechanism for Program Increment
creations so that we can reuse it.

Change-Id: I0c4085f5bc7ff36dc30139731e092ad049de5894