Gather Field values at given changeset of source PI
part of story #22675 replicate program increment changes to mirrored
program increments

How to test:
Make sure your redis worker is running

- Edit an existing Program Increment artifact and change its title.
- you should see a log message in
/var/log/tuleap/program_management_syslog:

[debug] Title value: <your updated title>

The log message only exists for testing purposes, to show that we
retrieve field values. It will be removed on next commits.

Notes:
I know the logic is very very similar between ReplicationData path and
ProgramIncrementUpdate path. Given the time constraints on the feature,
it seems preferable to introduce this duplication now and remove it when
we will work on Iterations. We should introduce a common interface
between the two, but that means changing (again) a lot of
ReplicationDataAdapter and it means extracting a wrapper event for
ArtifactCreated. Eventually, we will have to do this (and it will be a
good thing), but I think we can afford to do it later.

Change-Id: I74d906a0dffc74982df2ea5f1ab242ced130478d