feat: Display attachments below the comment when expanded
User can expand the comment zone in order to better read the comment and
see bigger images, or in order to have a comfortable size for the
editor.

Assuming that there will be not more than 5/7 attachments at max, if we
keep them displayed at the right of the comment when expanded, then a
lot of vertical space will be lost, thus reducing the advantage of
expanded mode, especially on smaller screens.

The proposal here is to display the attachments below the comment in
expanded mode. In collapsed mode, they are kept at the right of the
comment because we don't want to eat too much steps space.

Part of story #22705: display list of attachments of a test execution

Change-Id: I3b0747190268c4801795783e3b7f9daae31dda57