Collect error for plannable trackers
part of request #22583 Add missing configuration errors in program admin section

Why:
If configuration is incorrect you won't be able to see your artifact in backlog plannable list.
In artifact view there is a notice for user indicating that the element can not be added to top backlog,
This patch display the error for help administrator to configure the program

How to test:
Have a tracker without title and status semantic
Add it in plannable tracker list

Global warning message indicate user that configuration is incorrect

Please note, that plannable tracker are only concerned by this two semantics
all other errors corrected for PI or increment are meaningless in this context.

Change-Id: I636522cff0b5cd5ca06c6e4003946f5c84347d7c