feat: Allow parent artifact selection when project has a parent project
part of: story #22681 select a parent from Program for my features/enablers

Originally, the parent artifact selector was available only when the
current tracker had a parent tracker. In the SAFe context, we do not
have any tracker hierachy between a parent project and its children
trackers. In this case, we need to detect whether the current project is
a team, so we are able to display the select box when needed.

How to test:
- Given a tracker with no child relation to any other tracker of the
  current project
- Given that the current project is a child of another project in a program
  management context
- Given that the artifact has not a parent artifact yet
- Given that some artifacts are eligible to be parent in the parent
  project

--> In the artifact modal (kanban/PV2/TTM) the parent selector is displayed

Change-Id: I126369ee89840fad6b6f7e22f5a6e6f39d3eeb20