fix: Realtime issue when editing a comment in TTM
Two similar (more or less) issues are being fixed by this patch:

Scenario 1
----------
Given another user has marked as failed a test with a comment,
And I go to the test to see the reasons of the fail,
When another user change the status of another test,
Then the current execution comment does not disappear.

Scenario 2
----------
Given I am executing a test and this test fails
And I am writing a comment to explain the reasons of the fail,
When another user change the status of another test,
Then no "concurrent edit" warning should be displayed.

Part of story #22226: edit test execution comments

Change-Id: I8560865d1b9742af341c3a717d565af916f619fb