request #23036: A Git command executed in a repository forks 2 Git processes
A call to BranchName::defaultBranchName() do not generate an execution
of Git anymore. The default branch name is static so there is no need to
check it at runtime every time.
In the very unlikely situation where the current branch name becomes
invalid, we would still detect the issue early thanks to the added test.

Change-Id: Iaf5037cf15d19583e15f63c7578da3e7825a3367