Replace RichTextEditor by web component
part of request #22647 Convert Vue components to hybrids web components

How to test:
No functional change expected in the text fields or follow-up comments.
When you change the format, it will be saved when submitting. When you
change the content in the textarea (text format) or in CKEditor (html
format), it will be saved. You can still upload images by copy/pasting
or drag/dropping them. If your browser is offline (with dev tools),
there will be an error if you try to paste an image. If the image
exceeds the allowed size (edit your local.inc), there will be an error.
When there are no file fields, upload won't be allowed.

You can test basic non-regression with cypress end-to-end tests for
agiledashboard.

Change-Id: Ica2ab5422f088a50351140586bc3e9bd88377d62