request #23041: Use an absolute path when calling `gitolite access`
This make sure we do not encounter issue because gitolite is resolved to
a different path than the one defined in the sudoers file we deploy for
it. Using the absolute path also make sure we do not get unwanted
behaviors because system administrators have configured something for
their own needs in the sudoers secure_path setting.

Change-Id: Ie92e526d2ec38c0eb29e508bf26b40709081e530