Plugins can be installed/uninstalled on MySQL 8.0
The main goal of this contribution is to make DBTablesDaoLoadKnowDataUTF8MB4Test
passes the tests. Some migrations/upgrades code has been updated as well
because the issues were spotted at the same time.

The migration/upgrade files od the docman plugin have been removed. They
contained code not compatible with MySQL 8.0 and are not useful anymore.

Part of request #22660: Run Tuleap with MySQL 8.0

Change-Id: I7d350240951f0df7d8b94b9fd5f99258fff39a9d