Use ProgramIncrementCreation instead of ReplicationData
part of story #21368 create mirrored iteration

No functional change expected in the creation of Program Increment
mirrors. You can do the non-regression test with cypress end-to-end
tests.

This also renames "Runner" to "Dispatcher" to match the naming for PI
Updates.
In a following commit, the "Task" will be renamed to "Process" and will
also use ProgramIncrementCreation. ReplicationData will be removed.

Change-Id: Idd6e73747cee81f4c5751638e58313d1e2982f17