feat: Create/Delete/update project bot in project admin UI
This is part of story #22696 configure mattermost bots at project level

In the project admin menu, a new entry "Bot Mattermost" is added.
It allows to Create/Delete/update project bots, with the very same
limitations defined for system bots.

For now, a project bot cannot have the same configuration of an already
existing project or system bot with the same name + URL.

These bots can then be used in the Mattermost notifications of the
project.

Notes
-----

In newly created controllers, some errors lead to a rediction to the Tuleap homepage
with non translated error messages. This is done on purpose, because
these errors must not happen when using the UI (this can happen when
forging a request for instance).

Change-Id: Iff7b88ae0ecfcb623a25075fbb7eecd97bdf202b