Add the ability to choose a project icon
Part of story #22687 distinguish projects with an icon

The `emoji-button`[0] library has been used to have the icon picker
This libray has been chosen because it seems to be a library which is
still maintained and used.
It was also choosen because the lib is light and written in
VanillaJS/TS.
That lib have aslo some configurable options such as he possibility to
take in account translation or skin tone...

Since is not possible to remove a single emoji, a new data source has been
created without the unwanted emoji.

/!\ The feature is behind a feature flag

How to test:
 - In your bash web enable the feature => tuleap config-set feature_flag_project_icon_display 1
 - Go to the project admin Details menu.
 - Play with the icon picker
=> All the text should be translated.
 - Choose an icon
=> The chosen icon is displayed

Note: The chosen icon is not saved in backend yet

[0]: https://github.com/joeattardi/emoji-button

Change-Id: Ie8cab07d77199dad66a8810cd590b8b548cad8f7