fix: Preserve line breaks for default value in text fields
Given a text field with the following default value, that serves as a template to help users fill their artifacts:

```
As a…

I order to…

I want to…
```

Then the default value should stay the same, with line breaks,
regardless the preference value of text format for the user: text, html,
markdown. Before this contribution, in html, line breaks were lost which
resulted as having everything on one line.

No changes when editing an artifact with user's values.
If the request to submit/update the artifact is failing (missing
required field for example), then the submitted value is correctly
displayed.

Closes request #23371: Preserve line breaks for default value in text fields

Change-Id: I26950eb8632b4d84e19a30a6922fc69010bafe2c