chore(forge-config): Clean-up ForgeConfig loading
The main loading logic of the configuration was in pre.php without
proper test so the first thing was to transfer that logic inside
ForgeConfig so it can be controled and tested.

This change comes with removal, as much as possible, to specific config
loading (in some places we were loading only local.inc for instance) in
most of those places it was not needed to make a special case and, in
fine, might have provided incorrect values.

On ForgeConfig, `loadFromFile` was left public because it still used in
some tests (integration & e2e distlp) and it would have required too
much work to remove usage at this stage while there is a longer plan to
address the issue.

Part-of: story #22708 flexibility in the management of Tuleap Configuration

Change-Id: Id16e888c9ee2ce8f7ee32becb724026bdf81306c