fix: GraphOnTrackers error under PHP 8.0
We have two similar issues:
* When the report is not saved
* When an artifact does not have a value for the `start date` and/or
`due date` and/or `end date`

In that cases, there is an error displayed at the top of the report and
the gant chart is not displayed.

This is due to the saner string to number comparisons introduced in PHP
8.0. When an artifact has no start date, then the start date is an empty
string `''`. But when compared to 0 (to check if the date is set) [the
behavior is not the same between PHP 7 and PHP 8][rfc]:

var_dump('' == 0); // PHP 7: true. PHP 8: false.
https://3v4l.org/v4p2U

Closes request #23373: Gantt chart error when an artifact does not have a start date

[rfc]: https://wiki.php.net/rfc/string_to_number_comparison

Change-Id: I5a4c54b7d27bb90d5a18b2a2777f6518db31c41f