request #23372: Generated docx document from a tracker report cannot be opened in MS Word
The issue was caused by the fact the numbering generated for
the main titles had too many levels. After 9 levels Word (and Office Web
in write mode) refuses to open the files even if the levels are not used.

Anyway, since 6c6752dfd3aacea25ac71783d8bbae94299e54ef we only need one
level so we can simplify it.

There are other quirks with the numbering and Word but we are not
affected by them (yet). The most notable is that Word expects the
concrete numbering ID to start from 1 instead of 0. It only complains
about it when the concrete numbering with the ID 0 is used in the
document which is not the case for now.

Change-Id: I70a61d2edb874ee29e546b7c5950abcf7114b242