fix: More intuitive search in Mediawiki
In order to search in Mediawiki, one have to open the SwitchTo modal, to
enter terms, and to hit [Legacy search]. This is not really intuitive.

Mediawiki's search box is already present when MW is displayed in
Compatibility/Fullscreen mode (via Mediawiki service administration).
Enabling it for integrated mode is a cheap alternative to a standard,
full featured, global, consistent, top notch search in Tuleap that will
come later (no schedule yet).

Closes request #23376: More intuitive search in Mediawiki

Change-Id: I671b324dd649198a0537e22321881f9612fc061d