fix: Reopen fieldsets on artifact submission
Given a fieldset that contains an invalid field (e.g. required field
that has no value)
And this fieldset is collapsed
When user submit the form
Then nothing happen.

This is due because the native html5 validation wants to display a
message for an hidden input.

This patch fix the situation by automatically expand the fieldsets that
contains invalid inputs.

Closes request #23382: Invalid element in hidden fieldset prevents artifact submission

Change-Id: Ic6e8916838aa2521f42720df7da9481c03b5598e