request #23396: Use pnpm to replace npm and Lerna
The lockfile is not shared accross the workspace to avoid a dependencies
to external plugins.

In the Nix based environment the `npm` CLI is replaced by a shell script
exiting with an error code to avoid misusages.

A few dependencies are added in some packages. The way pnpm organizes
the node_modules folder forces us to explicitly define all our
dependencies.

angular-locker, dragular and ckeditor4 are no more transformed by Babel
when running the tests. Their format module combined with Babel, node
and pnpm does not seem to play well together. Anyway they do not need to be
transformed because they can be loaded in a node env directly. No
changes is made to the production build setup.

Change-Id: Iaa6b19b3489a34da3b5f64a363605376c25c2832