request #23379 I can bypass the creation check for Program increments
This patch only deal with Program Increment Iteration will be handle in dedicated patchset

Given I am a project member of the Program but I am not project member of one of the Teams, when I click on "New program increment", I don't see an error and I am redirected to artifact creation view.
I should have seen an error telling me I don't have permission to create a new program increment.

This issue also appears for Iterations (when active with the feature flag).

The logs show that the code cannot retrieve the top planning of the project I am not member of. The error log is written but the "submit" button is not disabled.

Change-Id: Ifb0da7c188c979bea58ff51257ee0f2dbecdc110