Kill ReplicationData 🔥
part of story #21368 create mirrored iteration

No functional change expected when creating or updating mirrors of
Program Increments. You can do the non-regression test with cypress
end-to-end tests.

ReplicationData is gone and replaced by ProgramIncrementCreation.
Creation of Program Increments is entirely handled by redis events, the
DB table is not used anymore.
Processing and Dispatching of the ProgramIncrementCreationEvent are now
separate (and easier to handle).

Change-Id: I7e3bfdb00313658ae9a43ffc6cff2e2cc19401f3