feat(ttm): Remove attachment files from an execution
This patch introduces an improvement of the REST route PUT testmanagement_executions/id
and insert a new optional key `deleted_file_ids` to remove files from an
execution.

For the moment, when the user gives a bad file id to delete, no exception
is thrown. That can be done in a future patch if necessary.

How to test:
 - You need to have an execution with attachment
 - In TTM, click on 'Edit comment'
 - You can click on the trash to mark the attachment as deleted
 - Click on "Save the comment"

Expected result:
 - There should be not the file attachment

This is part of story #22576 have multiple files uploaded with comments

Change-Id: I793a9119a0b92c2fd29ebae48fdb17c95acd1f42