request #23391 Enhance banners configuration UX
Save button is now disabled if the user tries to activate
the banner with an empty message. Before, the user was able
to click the button but it didn't do anything.

Moreover, having the message "No banner defined" even
when the user didn't save the configuration was misleading.
So the message has been removed.

Change-Id: I5beeb0f3536423949bcf43162d6e376e36a435d6