Use asynchronous event as the only source for Iteration creations
part of story #21368 create mirrored iteration

No functional change expected when creating mirrored iterations:
- Make sure your redis worker is running
- Run `tuleap config-set
feature_flag_program_management_display_iteration 1` to enable
iterations.
- Edit a Program Increment artifact and add one or more artifact links
with _is_child type to Iteration artifacts.
- a log message will be added to "program_management_syslog" for each
iteration:

[debug] Processing iteration creation with iteration #6318 for user

- Only links to Iterations that have never been linked in previous
changesets of the Program Increment are considered. If you edit the PI
again, the iterations are not processed again.

Notes:
This also introduces a "Builder" for the iteration creation processor
that follows in the footsteps of program increments. Later it will save
us memory on each worker event because we will only instantiate a big
stack of classes when needed.

Change-Id: Ibd467bffb9160bff0a1e20e3be4fe72a25ff664e