chore(forge-config)!: dump contains LDAP variables
This commit on its own is not really useful but it brings all the logic
so the next one can drop config file parsing in perl.

The core of this contribution is to replace the way plugin's
configuration files are loaded. Instead of having a custom parsing of
dot inc files, we re-use the logic already present in ForgeConfig. In
order to reduce the impact of the change, the existing API still allows
to get the variables from "PluginInfo".

In fact, without the future work on Variable loading, it's not actually
possible to use ForgeConfig directly to get variable values in Plugin.
It's because parsing of file still requires to pass via
'PluginInfo::loadPlugin' to work. While it's not perfect, the risk of
miss usage is low (the code that would get the variable from ForgeConfig
directly will not work) and it's a tradeoff to limit the amount of things to
change to get to the objective: remove external parsing of 'dot inc'
files.

BREAKING CHANGE: this commit might have an impact on old platforms (🙄) that
didn't follow the security recommendation of Tuleap 9 about removing the
edition of properties as Site Administrator. We will need to warn about
this in deployment guide & co.

Part-of: story #22708 flexibility in the management of Tuleap Configuration

Change-Id: I4d92332e9d81351539170d36bfc7ab295495eeff