Do not require UserReference in BuildProgram
part of story #21368 create mirrored iteration

No functional change expected.

UserIdentifier is easier to build than UserReference.
Almost all the time we have a UserIdentifier. Since we only need
UserReference for a single exception in BuildProgram, and given we must
retrieve the full \PFUser anyway, we can rebuild the UserReference at
this one place and relax the interface.

Change-Id: I1a86a2beff94a68ea38d3bed47bf76b933cab3d2