request #21420 jest: 26.6.3 -> 27.2.4
Release notes: https://jestjs.io/blog/2021/05/25/jest-27

ts-jest changelog:
https://github.com/kulshekhar/ts-jest/blob/v27.0.5/CHANGELOG.md

Packages using angular-mocks in tests are runned with jest-jasmine2
instead of the new jest-circus. Our usage of angular-mocks (or
angular-mocks itself?) does not seem to play well with jest-circus.

The test pipeline has been rewritten. It looks like, even with Jest 26,
we are hitting the node heap size limit. It looks like the issue is more
visible than before because when the worker dies Jest will not
re-schedule it anymore (which is probably a good thing in this context).
This is likely caused by memory leaks [0][1] which we should investigate but
since it is something we already experience it might be easier to
troubleshoot with the latest version of all the tools. In order to
workaround it, the heap size of node has been increased. To make sure
other tests still have enough memory to run the concurrency has been
reduced. The manual ordering of the test stages has been removed. No
stages run in sequence anymore but the concurrency is limited to 4
stages (the others will wait until one of the ongoing stage is done).
Ultimately it is probably a good thing because it leads to a better
usage of all the available CPU cores. The number of workers Jest uses
has been increased a bit since we reduced the pressure put on the CPUs.

[0] https://github.com/kulshekhar/ts-jest/issues/1967
[1] https://github.com/facebook/jest/issues/7874

Change-Id: I23811e43ea0cf6da4cf3fb92422d12f8a8e40894