fix: check invalid values for open list submission
Given you have a required open list at artifact submission
And you enter some text but do not select "Add …" to create a new value
Then the artifact is accepted and does not contain any value in the open list field.

Instead we should reject the artifact submission since it is a required
field. The current contribution only address the backend side. In order
to improve the UX, we could prevent the form submission on client side,
but it will come later, as an enhancement.

Part of request #23406: Required open list can be empty at artifact submission

Change-Id: Icf3dd118243d3c6ac8f041ca50ef54553b28661f