Split interface for milestone retrieval
part of story #21368 create mirrored iteration

No functional change expected in the "creation checks". There is still
an error message when your user cannot see one of the teams (because
it's private).

Single-method interfaces are easier to stub and to handle. We will need
the one for Mirrored Iteration Tracker.

Change-Id: Ie7eb6689ed4f4a08828f10fe1419bbb101103dcd