feat(forge-config): list keys that cannot be modified
Until now, ConfigListCommand was only showing variables that can be
modified with ConfigSetCommand. However, in a previous commit we
introduced a new configuration variable controlled by #[ConfigKey]
attribute but this attribute cannot be modified by ConfigSet. Unless you
are mad or a java developer, you are unlikely to store the info to
access your database INSIDE the very same database.

In the Future, ConfigListCommand will list all possible parameters and
ConfigSetCommand will allow to set most of those parameters. As some of
those parameters should not be set in the DB (like DB parameters) or
should not be modifiable with the unified API (like the number of
license in an external plugin) the #[CannotBeModified] attribute is
introduced.

Part-of: story #22708 flexibility in the management of Tuleap Configuration

Change-Id: Ia9a32217148f13babb54523545119977d6feeb49