Display the project icon in the breadcrumb of BP and FP pages
Part of story #22687 distinguish projects with an icon

This contribution concerns only the pages which does not use a Vue app to
build the breadcrumb.
Each services which use a Vue app to build the breadcrumb will have a
dedicated commit.

To test, you have to collapse the project sidebar

Example of page which can be tested:
 - First page of Tracker (BP)
 - Tracker report (FP)
 - First page Agile Dashboard (FP)
 - New Taskboard (BP)

Change-Id: Idead00b7e0259d32007e5888a537e7814917d8d6