request #23413: Workaround Jest/ts-jest/TypeScript memory leaks
To keep the leaks to an acceptable level we now use multiple Jest
instances to run the tests.

One Jest instance is spawn for every package having a test script.

Note that executing the tests in the tracker or agile dashboard plugin
directories does not execute the tests of their internal libraries. This
is needed to not play the same tests multiple times and it is more
consistent with the build process.

Change-Id: I96ed5f9a36bbfc797dfcb9201a20e9aeb884e1bd