Separate formatting from data objects
part of story #21368 create mirrored iteration

How to test:
No functional change expected in program increments or iterations
replication. You can do non-regression with cypress tests for program
increments.

Notes:
The "toFieldsDataArray()" method was strongly coupled with the Trackers
API, it was made for this purpose. Domain objects should not be tied to
other plugins, exactly as they should not be tied to REST JSON
representation or Presenters. Therefore, formatting is moved to the
Adapters namespace where it is called by the ArtifactCreator and
ChangesetAdder.

This allows us to "publish" the contents of
MirroredTimeboxChangesetValues, which lets us test them more deeply in
Processor tests. Processor tests are kind of integration tests, they
test that the Changesets produced at the end of the chain are built with
the correct data.

Change-Id: I56038ee33a963dd65701504e6b3d60a9cda6df63