Add support of HTML titles
The existing break we added at the end of each paragraph has been
removed. It does not exist in the DOM so there is no good reason to add
it and it causes excessive spacing with spacing.

Only smaller titles than the ones used to structure the document are used.
As a result we only have two levels, anything below (not including) <h1>
is considered as the heading level 6.

Part of story #22569: export `text` fields

Change-Id: I092e1d626602e148c3f13b7442ddd9ffbb004d28