Retrieve the artifact ID of the just created mirrored iterations
part of story #21368 create mirrored iteration

How to test:
- Make sure your redis worker is running
- Run `tuleap config-set
feature_flag_program_management_display_iteration 1` to enable
iterations.

1. If you have already made tests for this story, you can trigger
a redis event for an existing PI with Iterations linked to it:

- Enter the redis-cli in your redis container:

  $ docker exec -it tuleap_redis_1 redis-cli

- Push a message to the events queue manually:

> LPUSH 'app_user_events' '{"event_name":"tuleap.program_management.program_increment.update","payload":{"program_increment_id":6269,"user_id":110,"changeset_id":13408,"iterations":[{"id":6317,"changeset_id":13403},{"id":6318,"changeset_id":13404}]}}'

Change the ids above according to your setup. You can get changeset_id
in the Artifact View, there is a data-changeset-id attribute.

2. Otherwise, if you don't have pre-wired artifacts:

- Edit a Program Increment artifact and add one or more artifact links
with _is_child type to Iteration artifacts.

3. With either method 1. or 2., for each Team, new mirrored iteration
artifacts will be created and linked to the source iteration with
a "_mirrored_milestone" type. For each new mirrored iteration, there
will be a log message like:

  [debug] Created mirrored iteration 2098

The log message is temporary and will be removed in later commits.

Change-Id: I54dae06ef317d7646d7dc8b8852f58e42a5331d7